Reorganize structural model module
