+++ tests/statespace/models/structural/components/test_seasonality.py
@@ -1,0 +1,83 @@
+import numpy as np
+import pytest
+
+from pytensor import config
+
+from pymc_extras.statespace.models import structural as st
+from tests.statespace.models.structural.conftest import _assert_basic_coords_correct
+from tests.statespace.test_utilities import assert_pattern_repeats, simulate_from_numpy_model
+
+ATOL = 1e-8 if config.floatX.endswith("64") else 1e-4
+RTOL = 0 if config.floatX.endswith("64") else 1e-6
+
+
+@pytest.mark.parametrize("s", [10, 25, 50])
+@pytest.mark.parametrize("innovations", [True, False])
+@pytest.mark.parametrize("remove_first_state", [True, False])
+@pytest.mark.filterwarnings(
+    "ignore:divide by zero encountered in matmul:RuntimeWarning",
+    "ignore:overflow encountered in matmul:RuntimeWarning",
+    "ignore:invalid value encountered in matmul:RuntimeWarning",
+)
+def test_time_seasonality(s, innovations, remove_first_state, rng):
+    def random_word(rng):
+        return "".join(rng.choice(list("abcdefghijklmnopqrstuvwxyz")) for _ in range(5))
+
+    state_names = [random_word(rng) for _ in range(s)]
+    mod = st.TimeSeasonality(
+        season_length=s,
+        innovations=innovations,
+        name="season",
+        state_names=state_names,
+        remove_first_state=remove_first_state,
+    )
+    x0 = np.zeros(mod.k_states, dtype=config.floatX)
+    x0[0] = 1
+
+    params = {"season_coefs": x0}
+    if mod.innovations:
+        params["sigma_season"] = 0.0
+
+    x, y = simulate_from_numpy_model(mod, rng, params)
+    y = y.ravel()
+    if not innovations:
+        assert_pattern_repeats(y, s, atol=ATOL, rtol=RTOL)
+
+    # Check coords
+    mod.build(verbose=False)
+    _assert_basic_coords_correct(mod)
+    test_slice = slice(1, None) if remove_first_state else slice(None)
+    assert mod.coords["season_state"] == state_names[test_slice]
+
+
+def get_shift_factor(s):
+    s_str = str(s)
+    if "." not in s_str:
+        return 1
+    _, decimal = s_str.split(".")
+    return 10 ** len(decimal)
+
+
+@pytest.mark.parametrize("n", [*np.arange(1, 6, dtype="int").tolist(), None])
+@pytest.mark.parametrize("s", [5, 10, 25, 25.2])
+def test_frequency_seasonality(n, s, rng):
+    mod = st.FrequencySeasonality(season_length=s, n=n, name="season")
+    x0 = rng.normal(size=mod.n_coefs).astype(config.floatX)
+    params = {"season": x0, "sigma_season": 0.0}
+    k = get_shift_factor(s)
+    T = int(s * k)
+
+    x, y = simulate_from_numpy_model(mod, rng, params, steps=2 * T)
+    assert_pattern_repeats(y, T, atol=ATOL, rtol=RTOL)
+
+    # Check coords
+    mod.build(verbose=False)
+    _assert_basic_coords_correct(mod)
+    if n is None:
+        n = int(s // 2)
+    states = [f"season_{f}_{i}" for i in range(n) for f in ["Cos", "Sin"]]
+
+    # Remove the last state when the model is completely saturated
+    if s / n == 2.0:
+        states.pop()
+    assert mod.coords["season_state"] == states
