+++ tests/statespace/models/structural/test_against_statsmodels.py
@@ -4,15 +4,11 @@ import warnings
 from collections import defaultdict
 
 import numpy as np
-import pandas as pd
-import pymc as pm
 import pytensor
-import pytensor.tensor as pt
 import pytest
 import statsmodels.api as sm
 
 from numpy.testing import assert_allclose
-from scipy import linalg
 
 from pymc_extras.statespace import structural as st
 from pymc_extras.statespace.utils.constants import (
@@ -29,8 +25,6 @@ from tests.statespace.shared_fixtures import (  # pylint: disable=unused-import
     rng,
 )
 from tests.statespace.test_utilities import (
-    assert_pattern_repeats,
-    simulate_from_numpy_model,
     unpack_symbolic_matrices_with_params,
 )
 
@@ -104,15 +98,6 @@ def _assert_coord_shapes_match_matrices(mod, params):
         n_shocks,
         n_shocks,
     ), f"Q expected to have shape (n_shocks, n_shocks), found {Q.shape[-2:]}"
-
-
-def _assert_basic_coords_correct(mod):
-    assert mod.coords[ALL_STATE_DIM] == mod.state_names
-    assert mod.coords[ALL_STATE_AUX_DIM] == mod.state_names
-    assert mod.coords[SHOCK_DIM] == mod.shock_names
-    assert mod.coords[SHOCK_AUX_DIM] == mod.shock_names
-    assert mod.coords[OBS_STATE_DIM] == ["data"]
-    assert mod.coords[OBS_STATE_AUX_DIM] == ["data"]
 
 
 def _assert_keys_match(test_dict, expected_dict):
@@ -548,293 +533,3 @@ def test_structural_model_against_statsmodels(
     _assert_param_dims_correct(built_model.param_dims, expected_dims)
     _assert_coords_correct(built_model.coords, expected_coords)
     _assert_params_info_correct(built_model.param_info, built_model.coords, built_model.param_dims)
-
-
-def test_level_trend_model(rng):
-    mod = st.LevelTrendComponent(order=2, innovations_order=0)
-    params = {"initial_trend": [0.0, 1.0]}
-    x, y = simulate_from_numpy_model(mod, rng, params)
-
-    assert_allclose(np.diff(y), 1, atol=ATOL, rtol=RTOL)
-
-    # Check coords
-    mod = mod.build(verbose=False)
-    _assert_basic_coords_correct(mod)
-    assert mod.coords["trend_state"] == ["level", "trend"]
-
-
-def test_measurement_error(rng):
-    mod = st.MeasurementError("obs") + st.LevelTrendComponent(order=2)
-    mod = mod.build(verbose=False)
-
-    _assert_basic_coords_correct(mod)
-    assert "sigma_obs" in mod.param_names
-
-
-@pytest.mark.parametrize("order", [1, 2, [1, 0, 1]], ids=["AR1", "AR2", "AR(1,0,1)"])
-def test_autoregressive_model(order, rng):
-    ar = st.AutoregressiveComponent(order=order)
-    params = {
-        "ar_params": np.full((sum(ar.order),), 0.5, dtype=floatX),
-        "sigma_ar": 0.0,
-    }
-
-    x, y = simulate_from_numpy_model(ar, rng, params, steps=100)
-
-    # Check coords
-    ar.build(verbose=False)
-    _assert_basic_coords_correct(ar)
-    lags = np.arange(len(order) if isinstance(order, list) else order, dtype="int") + 1
-    if isinstance(order, list):
-        lags = lags[np.flatnonzero(order)]
-    assert_allclose(ar.coords["ar_lag"], lags)
-
-
-@pytest.mark.parametrize("s", [10, 25, 50])
-@pytest.mark.parametrize("innovations", [True, False])
-@pytest.mark.parametrize("remove_first_state", [True, False])
-@pytest.mark.filterwarnings(
-    "ignore:divide by zero encountered in matmul:RuntimeWarning",
-    "ignore:overflow encountered in matmul:RuntimeWarning",
-    "ignore:invalid value encountered in matmul:RuntimeWarning",
-)
-def test_time_seasonality(s, innovations, remove_first_state, rng):
-    def random_word(rng):
-        return "".join(rng.choice(list("abcdefghijklmnopqrstuvwxyz")) for _ in range(5))
-
-    state_names = [random_word(rng) for _ in range(s)]
-    mod = st.TimeSeasonality(
-        season_length=s,
-        innovations=innovations,
-        name="season",
-        state_names=state_names,
-        remove_first_state=remove_first_state,
-    )
-    x0 = np.zeros(mod.k_states, dtype=floatX)
-    x0[0] = 1
-
-    params = {"season_coefs": x0}
-    if mod.innovations:
-        params["sigma_season"] = 0.0
-
-    x, y = simulate_from_numpy_model(mod, rng, params)
-    y = y.ravel()
-    if not innovations:
-        assert_pattern_repeats(y, s, atol=ATOL, rtol=RTOL)
-
-    # Check coords
-    mod.build(verbose=False)
-    _assert_basic_coords_correct(mod)
-    test_slice = slice(1, None) if remove_first_state else slice(None)
-    assert mod.coords["season_state"] == state_names[test_slice]
-
-
-def get_shift_factor(s):
-    s_str = str(s)
-    if "." not in s_str:
-        return 1
-    _, decimal = s_str.split(".")
-    return 10 ** len(decimal)
-
-
-@pytest.mark.parametrize("n", [*np.arange(1, 6, dtype="int").tolist(), None])
-@pytest.mark.parametrize("s", [5, 10, 25, 25.2])
-def test_frequency_seasonality(n, s, rng):
-    mod = st.FrequencySeasonality(season_length=s, n=n, name="season")
-    x0 = rng.normal(size=mod.n_coefs).astype(floatX)
-    params = {"season": x0, "sigma_season": 0.0}
-    k = get_shift_factor(s)
-    T = int(s * k)
-
-    x, y = simulate_from_numpy_model(mod, rng, params, steps=2 * T)
-    assert_pattern_repeats(y, T, atol=ATOL, rtol=RTOL)
-
-    # Check coords
-    mod.build(verbose=False)
-    _assert_basic_coords_correct(mod)
-    if n is None:
-        n = int(s // 2)
-    states = [f"season_{f}_{i}" for i in range(n) for f in ["Cos", "Sin"]]
-
-    # Remove the last state when the model is completely saturated
-    if s / n == 2.0:
-        states.pop()
-    assert mod.coords["season_state"] == states
-
-
-cycle_test_vals = zip([None, None, 3, 5, 10], [False, True, True, False, False])
-
-
-def test_cycle_component_deterministic(rng):
-    cycle = st.CycleComponent(
-        name="cycle", cycle_length=12, estimate_cycle_length=False, innovations=False
-    )
-    params = {"cycle": np.array([1.0, 1.0], dtype=floatX)}
-    x, y = simulate_from_numpy_model(cycle, rng, params, steps=12 * 12)
-
-    assert_pattern_repeats(y, 12, atol=ATOL, rtol=RTOL)
-
-
-def test_cycle_component_with_dampening(rng):
-    cycle = st.CycleComponent(
-        name="cycle", cycle_length=12, estimate_cycle_length=False, innovations=False, dampen=True
-    )
-    params = {"cycle": np.array([10.0, 10.0], dtype=floatX), "cycle_dampening_factor": 0.75}
-    x, y = simulate_from_numpy_model(cycle, rng, params, steps=100)
-
-    # Check that the cycle dampens to zero over time
-    assert_allclose(y[-1], 0.0, atol=ATOL, rtol=RTOL)
-
-
-def test_cycle_component_with_innovations_and_cycle_length(rng):
-    cycle = st.CycleComponent(
-        name="cycle", estimate_cycle_length=True, innovations=True, dampen=True
-    )
-    params = {
-        "cycle": np.array([1.0, 1.0], dtype=floatX),
-        "cycle_length": 12.0,
-        "cycle_dampening_factor": 0.95,
-        "sigma_cycle": 1.0,
-    }
-
-    x, y = simulate_from_numpy_model(cycle, rng, params)
-
-    cycle.build(verbose=False)
-    _assert_basic_coords_correct(cycle)
-
-
-def test_exogenous_component(rng):
-    data = rng.normal(size=(100, 2)).astype(floatX)
-    mod = st.RegressionComponent(state_names=["feature_1", "feature_2"], name="exog")
-
-    params = {"beta_exog": np.array([1.0, 2.0], dtype=floatX)}
-    exog_data = {"data_exog": data}
-    x, y = simulate_from_numpy_model(mod, rng, params, exog_data)
-
-    # Check that the generated data is just a linear regression
-    assert_allclose(y, data @ params["beta_exog"], atol=ATOL, rtol=RTOL)
-
-    mod.build(verbose=False)
-    _assert_basic_coords_correct(mod)
-    assert mod.coords["exog_state"] == ["feature_1", "feature_2"]
-
-
-def test_adding_exogenous_component(rng):
-    data = rng.normal(size=(100, 2)).astype(floatX)
-    reg = st.RegressionComponent(state_names=["a", "b"], name="exog")
-    ll = st.LevelTrendComponent(name="level")
-
-    seasonal = st.FrequencySeasonality(name="annual", season_length=12, n=4)
-    mod = reg + ll + seasonal
-
-    assert mod.ssm["design"].eval({"data_exog": data}).shape == (100, 1, 2 + 2 + 8)
-    assert_allclose(mod.ssm["design", 5, 0, :2].eval({"data_exog": data}), data[5])
-
-
-def test_add_components():
-    ll = st.LevelTrendComponent(order=2)
-    se = st.TimeSeasonality(name="seasonal", season_length=12)
-    mod = ll + se
-
-    ll_params = {
-        "initial_trend": np.zeros(2, dtype=floatX),
-        "sigma_trend": np.ones(2, dtype=floatX),
-    }
-    se_params = {
-        "seasonal_coefs": np.ones(11, dtype=floatX),
-        "sigma_seasonal": 1.0,
-    }
-    all_params = ll_params.copy()
-    all_params.update(se_params)
-
-    (ll_x0, ll_P0, ll_c, ll_d, ll_T, ll_Z, ll_R, ll_H, ll_Q) = unpack_symbolic_matrices_with_params(
-        ll, ll_params
-    )
-    (se_x0, se_P0, se_c, se_d, se_T, se_Z, se_R, se_H, se_Q) = unpack_symbolic_matrices_with_params(
-        se, se_params
-    )
-    x0, P0, c, d, T, Z, R, H, Q = unpack_symbolic_matrices_with_params(mod, all_params)
-
-    for property in ["param_names", "shock_names", "param_info", "coords", "param_dims"]:
-        assert [x in getattr(mod, property) for x in getattr(ll, property)]
-        assert [x in getattr(mod, property) for x in getattr(se, property)]
-
-    ll_mats = [ll_T, ll_R, ll_Q]
-    se_mats = [se_T, se_R, se_Q]
-    all_mats = [T, R, Q]
-
-    for ll_mat, se_mat, all_mat in zip(ll_mats, se_mats, all_mats):
-        assert_allclose(all_mat, linalg.block_diag(ll_mat, se_mat), atol=ATOL, rtol=RTOL)
-
-    ll_mats = [ll_x0, ll_c, ll_Z]
-    se_mats = [se_x0, se_c, se_Z]
-    all_mats = [x0, c, Z]
-    axes = [0, 0, 1]
-
-    for ll_mat, se_mat, all_mat, axis in zip(ll_mats, se_mats, all_mats, axes):
-        assert_allclose(all_mat, np.concatenate([ll_mat, se_mat], axis=axis), atol=ATOL, rtol=RTOL)
-
-
-def test_filter_scans_time_varying_design_matrix(rng):
-    time_idx = pd.date_range(start="2000-01-01", freq="D", periods=100)
-    data = pd.DataFrame(rng.normal(size=(100, 2)), columns=["a", "b"], index=time_idx)
-
-    y = pd.DataFrame(rng.normal(size=(100, 1)), columns=["data"], index=time_idx)
-
-    reg = st.RegressionComponent(state_names=["a", "b"], name="exog")
-    mod = reg.build(verbose=False)
-
-    with pm.Model(coords=mod.coords) as m:
-        data_exog = pm.Data("data_exog", data.values)
-
-        x0 = pm.Normal("x0", dims=["state"])
-        P0 = pm.Deterministic("P0", pt.eye(mod.k_states), dims=["state", "state_aux"])
-        beta_exog = pm.Normal("beta_exog", dims=["exog_state"])
-
-        mod.build_statespace_graph(y)
-        x0, P0, c, d, T, Z, R, H, Q = mod.unpack_statespace()
-        pm.Deterministic("Z", Z)
-
-        prior = pm.sample_prior_predictive(draws=10)
-
-    prior_Z = prior.prior.Z.values
-    assert prior_Z.shape == (1, 10, 100, 1, 2)
-    assert_allclose(prior_Z[0, :, :, 0, :], data.values[None].repeat(10, axis=0))
-
-
-@pytest.mark.skipif(floatX.endswith("32"), reason="Prior covariance not PSD at half-precision")
-def test_extract_components_from_idata(rng):
-    time_idx = pd.date_range(start="2000-01-01", freq="D", periods=100)
-    data = pd.DataFrame(rng.normal(size=(100, 2)), columns=["a", "b"], index=time_idx)
-
-    y = pd.DataFrame(rng.normal(size=(100, 1)), columns=["data"], index=time_idx)
-
-    ll = st.LevelTrendComponent()
-    season = st.FrequencySeasonality(name="seasonal", season_length=12, n=2, innovations=False)
-    reg = st.RegressionComponent(state_names=["a", "b"], name="exog")
-    me = st.MeasurementError("obs")
-    mod = (ll + season + reg + me).build(verbose=False)
-
-    with pm.Model(coords=mod.coords) as m:
-        data_exog = pm.Data("data_exog", data.values)
-
-        x0 = pm.Normal("x0", dims=["state"])
-        P0 = pm.Deterministic("P0", pt.eye(mod.k_states), dims=["state", "state_aux"])
-        beta_exog = pm.Normal("beta_exog", dims=["exog_state"])
-        initial_trend = pm.Normal("initial_trend", dims=["trend_state"])
-        sigma_trend = pm.Exponential("sigma_trend", 1, dims=["trend_shock"])
-        seasonal_coefs = pm.Normal("seasonal", dims=["seasonal_state"])
-        sigma_obs = pm.Exponential("sigma_obs", 1)
-
-        mod.build_statespace_graph(y)
-
-        x0, P0, c, d, T, Z, R, H, Q = mod.unpack_statespace()
-        prior = pm.sample_prior_predictive(draws=10)
-
-    filter_prior = mod.sample_conditional_prior(prior)
-    comp_prior = mod.extract_components_from_idata(filter_prior)
-    comp_states = comp_prior.filtered_prior.coords["state"].values
-    expected_states = ["LevelTrend[level]", "LevelTrend[trend]", "seasonal", "exog[a]", "exog[b]"]
-    missing = set(comp_states) - set(expected_states)
-
-    assert len(missing) == 0, missing
